rename render.hpp to renderer.hpp
