+++ include/engine/dirkengine.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "core/globals.hpp"
-#include "render/render.hpp"
+#include "render/renderer.hpp"
 
 DECLARE_LOG_CATEGORY_EXTERN(LogEngine)
